requirements.txt
django==3.1.4
django-allauth==0.44.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
django-dynamic-preferences==1.10.1
drf-api-tracking==1.7.8
django_multitenant==2.3.2
django-filter==2.4.0
django-reversion==3.0.8
django-cors-headers==3.5.0
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-storages==1.10.1
django-model-utils==4.1.1
django-dbbackup==3.3.0
bootstrap-admin==0.4.4
django-sendsms==0.4
django-anymail[mailgun]==8.1
django-phonenumber-field==5.0.0
django-ckeditor==6.0.0
jsonfield==3.1.0
redis==3.5.3
phonenumbers==8.12.14
simplejson==3.17.2
psycopg2-binary==2.8.6
python-dateutil==2.8.1
requests==2.25.0
twilio==6.49.0
isodate==0.6.0
pillow==8.0.1
boto3==1.16.30
awscli==1.18.190
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
stripe==2.55.1
cerberus==1.3.2